Minor cleanups
Restructured file directories so that each Design has it's own subdirectory, added package declarations for each directory
