+++ Design6/PointCPd6.java
@@ -1,7 +1,7 @@
 // This file contains material supporting section 2.9 of the textbook:
 // "Object Oriented Software Engineering" and is issued under the open-source
 // license found at www.lloseng.com 
-
+package Design6;
 /**
  * This class contains instances of coordinates in either polar or
  * cartesian format.  It also provides the utilities to convert
@@ -12,7 +12,7 @@
  * @author Dr Timothy C. Lethbridge
  * @version July 2000
  */
-public class PointCPd2
+public class PointCPd6
 {
   //Instance variables ************************************************
 
@@ -40,7 +40,7 @@ public class PointCPd2
   /**
    * Constructs a coordinate object, with a type identifier.
    */
-  public PointCPd2(char type, double xOrRho, double yOrTheta)
+  public PointCP6(char type, double xOrRho, double yOrTheta)
   {
     if(type != 'C' && type != 'P')
       throw new IllegalArgumentException();
@@ -84,13 +84,38 @@ public class PointCPd2
     else 
       return Math.toDegrees(Math.atan2(yOrTheta, xOrRho));
   }
+  
+	
+  /**
+   * Converts Cartesian coordinates to Polar coordinates.
+   */
+  public void convertStorageToPolar()
+  {
+    if(typeCoord != 'P')
+    {
+      //Calculate RHO and THETA
+      double temp = getRho();
+      yOrTheta = getTheta();
+      xOrRho = temp;
+      
+      typeCoord = 'P';  //Change coord type identifier
+    }
+  }
 	
   /**
    * Converts Polar coordinates to Cartesian coordinates.
    */
-  public String getCartesian()
+  public void convertStorageToCartesian()
   {
-    return ("("+getX()+", "+getY());
+    if(typeCoord != 'C')
+    {
+      //Calculate X and Y
+      double temp = getX();
+      yOrTheta = getY();
+      xOrRho = temp;
+   
+      typeCoord = 'C';	//Change coord type identifier
+    }
   }
 
   /**
@@ -101,7 +126,7 @@ public class PointCPd2
    * @param pointB The second point.
    * @return The distance between the two points.
    */
-  public double getDistance(PointCPd2 pointB)
+  public double getDistance(PointCP pointB)
   {
     // Obtain differences in X and Y, sign is not important as these values
     // will be squared later.
@@ -119,13 +144,13 @@ public class PointCPd2
    * @param rotation The number of degrees to rotate the point.
    * @return The rotated image of the original point.
    */
-  public PointCPd2 rotatePoint(double rotation)
+  public PointCP rotatePoint(double rotation)
   {
     double radRotation = Math.toRadians(rotation);
     double X = getX();
     double Y = getY();
         
-    return new PointCPd2('C',
+    return new PointCP('C',
       (Math.cos(radRotation) * X) - (Math.sin(radRotation) * Y),
       (Math.sin(radRotation) * X) + (Math.cos(radRotation) * Y));
   }
